requirements.txt
open3d>=0.18.0
trimesh>=4.0.0
numpy>=1.24.0
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
python-multipart>=0.0.9
openai>=1.30.0
python-dotenv>=1.0.0
matplotlib>=3.8.0
